（大塚）SpotRegisterウィンドウのバグ修正
修正
・GoogleMapのURLから緯度経度を抽出する際に、
場所自体の緯度経度ではなく、そのときGoogleMapで見ていた場所の中心緯度経度が
抽出されてしまう問題の修正

+++ Assets/Editor/SpotDataRegister/SpotRegisterBody/CreateSpotAsset.cs
@@ -68,8 +68,8 @@ public class CreateSpotAsset
             if (!isGoogleMapUrl) throw new InVaildGoogleMapURL();
 
             bool successful = false;
-            float tLong;
-            float tLet;
+            double tLong;
+            double tLet;
 
             successful = GoogleMapUrlLibrary.ExtractRetAndLetByUrl(registerDatas.spotUrl,out tLong,out tLet);
